Create UpdatedCommonHeader
The "common.h" file that comes with the HTSPB has not been updated for new changes in the latest version of RobotC.
Many functions have been depreciated and others do not work due to new "pointer" features.
FTC Team 8398: The Big Bot Theory has fixed all of these errors and below is the updated code! Enjoy!